Rename copy of RunAppBuilder to SDKMethodRunner
